Refs #3868. Clear compiler warnings.

+++ Code/Mantid/MantidPlot/src/Mantid/MantidUI.cpp
@@ -2965,7 +2965,7 @@ void MantidUI::showSequentialPlot(Ui::SequentialFitDialog* ui, MantidQt::MantidW
     Mantid::API::AnalysisDataService::Instance().retrieve(wsName) );
   if (ws)
   {
-    if ((ws->columnCount() - 1)/2 != (int)fitbrowser->compositeFunction()->nParams()) return;
+    if ((ws->columnCount() - 1)/2 != fitbrowser->compositeFunction()->nParams()) return;
     Table *t = importTableWorkspace(QString::fromStdString(wsName));
     if (!t) return;
     QString parName;
